# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cria um novo usuario",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf",
                                    "password",
                                    "password_confirmation",
                                    "telefone",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "rodrigo lindo"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "14785236945"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "@saudell123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "@saudell123"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "5544978947894"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "rodrigolindo@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "password"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "11122233347"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login realizado com sucesso"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/esquecisenha": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Esqueci minha senha",
                "operationId": "74436bbc973044f71b293ba6f2ad1c66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "11122233347"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Senha alterada com sucesso"
                    },
                    "404": {
                        "description": "CPF não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}